Complete code restructure
